Фаза 3: Слить core/types/ в core/types.py
