--- core/types/__init__.py
@@ -1,28 +1,0 @@
-"""Пакет типов домена для Clean Code рефакторинга."""
-
-from core.types._base import (
-    CharacterClass,
-    EquippedState,
-    GameDifficulty,
-    InventoryItem,
-    LanguageCode,
-    RuntimeSettings,
-    StatMap,
-    StringsDict,
-)
-from core.types.character_params import CharacterBuildParams
-from core.types.proficiencies import Expertise, Proficiencies
-
-__all__ = [
-    "CharacterBuildParams",
-    "CharacterClass",
-    "EquippedState",
-    "Expertise",
-    "GameDifficulty",
-    "InventoryItem",
-    "LanguageCode",
-    "Proficiencies",
-    "RuntimeSettings",
-    "StatMap",
-    "StringsDict",
-]
